requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
colorama==0.4.4
cycler==0.10.0
Django==3.2.5
django-phonenumber-field==5.2.0
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
phonenumbers==8.12.28
Pillow==8.3.1
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
